Recoded using RE and  BeautifulSoup
I retooled this API using BeautifulSoup and RE as etree was breaking
down when I ran it.   Right now, there are no controls other than
re-initializing the cache.   There are many redundancies in this code,
and I am using files, more than running in memory due to the issues I
encountered in etree.  I do plan on adding back some controls so this
can be executed with a single function call.  I also tried and failed
to reduce my match calls to a single function.  This code has been
tested and it works.  So, I will probably leave this all as is unless
my OCD starts up.